+++ examples/run_examples.jl
@@ -27,7 +27,7 @@ decompose_gates = ["decompose_hadamard",
                    "decompose_iSwapGate",
                    "decompose_RX_on_q3"]
 
-# decompose_gates = ["decompose_magic_M"]
+decompose_gates = ["decompose_magic_M"]
 
 #----------------------------------------------#
 #      Quantum Circuit Optimization model      #
@@ -39,6 +39,7 @@ for gates in decompose_gates
     qcm_optimizer = get_solver(params)
     global result_qc = QCO.run_QCModel(params, 
                                        qcm_optimizer, 
-                                       model_type = "compact_formulation")
+                                       model_type = "compact_formulation",
+                                       convex_hull_complex_gate_constraints = true)
 end