+++ src/constraints.jl
@@ -2,7 +2,7 @@
 # Build all constraints of the QuantumCircuitModel here  #
 #--------------------------------------------------------#
 
-# import QHull: chull 
+import QHull: chull
 
 function constraint_single_gate_per_depth(qcm::QuantumCircuitModel)
 
@@ -289,149 +289,149 @@ end
 
 function constraint_convex_hull_complex_gates(qcm::QuantumCircuitModel)
 
-#     if !qcm.data["are_gates_real"] 
+    if !qcm.data["are_gates_real"] 
 
-#         max_ex_pt = 10 # (>= 2) A parameter which can be an user input
+        max_ex_pt = 10 # (>= 2) A parameter which can be an user input
 
-#         z_onoff_var = qcm.variables[:z_onoff_var]
+        z_onoff_var = qcm.variables[:z_onoff_var]
 
-#         gates_real = qcm.data["gates_real"]
-#         gates_dict = qcm.data["gates_dict"]
+        gates_real = qcm.data["gates_real"]
+        gates_dict = qcm.data["gates_dict"]
 
-#         num_gates = size(gates_real)[3]
-#         depth     = qcm.data["depth"]
-#         n_r       = size(gates_dict["1"]["matrix"])[1]
-#         n_c       = size(gates_dict["1"]["matrix"])[2]
+        num_gates = size(gates_real)[3]
+        depth     = qcm.data["depth"]
+        n_r       = size(gates_dict["1"]["matrix"])[1]
+        n_c       = size(gates_dict["1"]["matrix"])[2]
 
-#         num_facets = 0
+        num_facets = 0
 
-#         for I=1:n_r
-#             for J=1:n_c
+        for I=1:n_r
+            for J=1:n_c
                 
-#                 vertices_coord = Set()
+                vertices_coord = Set()
 
-#                 for K in keys(gates_dict)
+                for K in keys(gates_dict)
 
-#                     re = QCO.round_real_value(real(gates_dict[K]["matrix"][I,J]))
-#                     im = QCO.round_real_value(imag(gates_dict[K]["matrix"][I,J]))
+                    re = QCO.round_real_value(real(gates_dict[K]["matrix"][I,J]))
+                    im = QCO.round_real_value(imag(gates_dict[K]["matrix"][I,J]))
 
-#                     push!(vertices_coord, (re, im))
-#                 end
+                    push!(vertices_coord, (re, im))
+                end
 
-#                 if (isapprox(minimum([x[1] for x in vertices_coord]), maximum([x[1] for x in vertices_coord]), atol = 1E-6)) || (isapprox(minimum([x[2] for x in vertices_coord]), maximum([x[2] for x in vertices_coord]), atol = 1E-6))
-#                     continue
-#                 end
+                if (isapprox(minimum([x[1] for x in vertices_coord]), maximum([x[1] for x in vertices_coord]), atol = 1E-6)) || (isapprox(minimum([x[2] for x in vertices_coord]), maximum([x[2] for x in vertices_coord]), atol = 1E-6))
+                    continue
+                end
 
-#                 if length(vertices_coord) == 2 
+                if length(vertices_coord) == 2 
                     
-#                     vertices = Vector{Vector{Float64}}()
+                    vertices = Vector{Vector{Float64}}()
 
-#                     for l in vertices_coord
-#                         push!(vertices, [l[1], l[2]])
-#                     end
+                    for l in vertices_coord
+                        push!(vertices, [l[1], l[2]])
+                    end
 
-#                     m,c = QCO._get_constraint_slope_intercept(vertices[1], vertices[2])
+                    m,c = QCO._get_constraint_slope_intercept(vertices[1], vertices[2])
                     
-#                     if isapprox(abs(m), 0, atol=1E-6)
+                    if isapprox(abs(m), 0, atol=1E-6)
 
-#                         JuMP.@constraint(qcm.model, [d=1:depth], 
-#                                          sum(gates_real[(2*I-1),(2*J), n_g] * z_onoff_var[n_g,d] for n_g = 1:num_gates) - c == 0)
-#                     else
+                        JuMP.@constraint(qcm.model, [d=1:depth], 
+                                         sum(gates_real[(2*I-1),(2*J), n_g] * z_onoff_var[n_g,d] for n_g = 1:num_gates) - c == 0)
+                    else
                         
-#                         JuMP.@constraint(qcm.model, [d=1:depth], sum(gates_real[(2*I-1),(2*J), n_g] * z_onoff_var[n_g,d] for n_g = 1:num_gates) 
-#                                                                  - slope*sum(gates_real[(2*I-1),(2*J-1), n_g] * z_onoff_var[n_g,d] for n_g = 1:num_gates) - c == 0)
-#                     end
-#                     num_facets += 1
+                        JuMP.@constraint(qcm.model, [d=1:depth], sum(gates_real[(2*I-1),(2*J), n_g] * z_onoff_var[n_g,d] for n_g = 1:num_gates) 
+                                                                 - slope*sum(gates_real[(2*I-1),(2*J-1), n_g] * z_onoff_var[n_g,d] for n_g = 1:num_gates) - c == 0)
+                    end
+                    num_facets += 1
 
-#                 elseif (length(vertices_coord) > 2) && (length(vertices_coord) <= max_ex_pt)
+                elseif (length(vertices_coord) > 2) && (length(vertices_coord) <= max_ex_pt)
                 
-#                     vertices = zeros(length(vertices_coord),2)
+                    vertices = zeros(length(vertices_coord),2)
 
-#                     ctr = 1
-#                     for l in vertices_coord
-#                         vertices[ctr,:] = [l[1] l[2]]
-#                         ctr += 1
-#                     end
+                    ctr = 1
+                    for l in vertices_coord
+                        vertices[ctr,:] = [l[1] l[2]]
+                        ctr += 1
+                    end
                     
-#                     # vertices_convex_hull = Polyhedra.planar_hull(Polyhedra.vrep(vertices)).points.points
-#                     vertices_convex_hull = vertices[QHull.chull(vertices).vertices, :]
-#                     num_ex_pt = size(vertices_convex_hull)[1]
+                    # vertices_convex_hull = Polyhedra.planar_hull(Polyhedra.vrep(vertices)).points.points
+                    vertices_convex_hull = vertices[QHull.chull(vertices).vertices, :]
+                    num_ex_pt = size(vertices_convex_hull)[1]
 
-#                     for i=1:num_ex_pt
-#                         v1 = vertices_convex_hull[i,:]
+                    for i=1:num_ex_pt
+                        v1 = vertices_convex_hull[i,:]
 
-#                         if i == num_ex_pt
-#                             v2 = vertices_convex_hull[1,:]
-#                         else 
-#                             v2 = vertices_convex_hull[i+1,:]
-#                         end
+                        if i == num_ex_pt
+                            v2 = vertices_convex_hull[1,:]
+                        else 
+                            v2 = vertices_convex_hull[i+1,:]
+                        end
 
-#                         # Test-vertex for half-space directionality
-#                         if i == (num_ex_pt - 1)
-#                             v3 = vertices_convex_hull[1,:]
-#                         elseif i == num_ex_pt
-#                             v3 = vertices_convex_hull[2,:]
-#                         else 
-#                             v3 = vertices_convex_hull[i+2,:]
-#                         end
+                        # Test-vertex for half-space directionality
+                        if i == (num_ex_pt - 1)
+                            v3 = vertices_convex_hull[1,:]
+                        elseif i == num_ex_pt
+                            v3 = vertices_convex_hull[2,:]
+                        else 
+                            v3 = vertices_convex_hull[i+2,:]
+                        end
 
-#                         slope, intercept = QCO._get_constraint_slope_intercept(v1, v2)
+                        slope, intercept = QCO._get_constraint_slope_intercept(v1, v2)
                         
-#                         # Facets of the hull
-#                         if !isinf(slope)
+                        # Facets of the hull
+                        if !isinf(slope)
 
-#                             if v3[2] - slope*v3[1] - intercept <= -1E-6
+                            if v3[2] - slope*v3[1] - intercept <= -1E-6
 
-#                                 if isapprox(abs(slope), 0, atol=1E-6)
+                                if isapprox(abs(slope), 0, atol=1E-6)
                                     
-#                                     JuMP.@constraint(qcm.model, [d=1:depth], sum(gates_real[(2*I-1),(2*J), n_g] * z_onoff_var[n_g,d] for n_g = 1:num_gates) - intercept <= 0)
-#                                 else
+                                    JuMP.@constraint(qcm.model, [d=1:depth], sum(gates_real[(2*I-1),(2*J), n_g] * z_onoff_var[n_g,d] for n_g = 1:num_gates) - intercept <= 0)
+                                else
                                     
-#                                     JuMP.@constraint(qcm.model, [d=1:depth], sum(gates_real[(2*I-1),(2*J), n_g] * z_onoff_var[n_g,d] for n_g = 1:num_gates) 
-#                                                                         - slope*(sum(gates_real[(2*I-1),(2*J-1), n_g] * z_onoff_var[n_g,d] for n_g = 1:num_gates)) - intercept <= 0)
-#                                 end
-#                                 num_facets += 1
+                                    JuMP.@constraint(qcm.model, [d=1:depth], sum(gates_real[(2*I-1),(2*J), n_g] * z_onoff_var[n_g,d] for n_g = 1:num_gates) 
+                                                                        - slope*(sum(gates_real[(2*I-1),(2*J-1), n_g] * z_onoff_var[n_g,d] for n_g = 1:num_gates)) - intercept <= 0)
+                                end
+                                num_facets += 1
 
-#                             elseif v3[2] - slope*v3[1] - intercept >= 1E-6
+                            elseif v3[2] - slope*v3[1] - intercept >= 1E-6
 
-#                                 if isapprox(abs(slope), 0, atol=1E-6)
+                                if isapprox(abs(slope), 0, atol=1E-6)
                                     
-#                                     JuMP.@constraint(qcm.model, [d=1:depth], sum(gates_real[(2*I-1),(2*J), n_g] * z_onoff_var[n_g,d] for n_g = 1:num_gates) - intercept >= 0)
-#                                 else
+                                    JuMP.@constraint(qcm.model, [d=1:depth], sum(gates_real[(2*I-1),(2*J), n_g] * z_onoff_var[n_g,d] for n_g = 1:num_gates) - intercept >= 0)
+                                else
                                     
-#                                     JuMP.@constraint(qcm.model, [d=1:depth], sum(gates_real[(2*I-1),(2*J), n_g] * z_onoff_var[n_g,d] for n_g = 1:num_gates) 
-#                                                                         - slope*(sum(gates_real[(2*I-1),(2*J-1), n_g] * z_onoff_var[n_g,d] for n_g = 1:num_gates)) - intercept >= 0)
-#                                 end
-#                                 num_facets += 1
+                                    JuMP.@constraint(qcm.model, [d=1:depth], sum(gates_real[(2*I-1),(2*J), n_g] * z_onoff_var[n_g,d] for n_g = 1:num_gates) 
+                                                                        - slope*(sum(gates_real[(2*I-1),(2*J-1), n_g] * z_onoff_var[n_g,d] for n_g = 1:num_gates)) - intercept >= 0)
+                                end
+                                num_facets += 1
                                 
-#                             else 
-#                                 Memento.warn(_LOGGER, "Indeterminate direction for the planar-hull cut")
-#                             end
+                            else 
+                                Memento.warn(_LOGGER, "Indeterminate direction for the planar-hull cut")
+                            end
 
-#                         else isinf(slope)
+                        else isinf(slope)
 
-#                             if v3[1] >= v1[1] + 1E-6
+                            if v3[1] >= v1[1] + 1E-6
                                 
-#                                 JuMP.@constraint(qcm.model, [d=1:depth], sum(gates_real[(2*I-1),(2*J-1), n_g] * z_onoff_var[n_g,d] for n_g = 1:num_gates) >= v1[1])
-#                             elseif v3[1] <= v1[1] - 1E-6
+                                JuMP.@constraint(qcm.model, [d=1:depth], sum(gates_real[(2*I-1),(2*J-1), n_g] * z_onoff_var[n_g,d] for n_g = 1:num_gates) >= v1[1])
+                            elseif v3[1] <= v1[1] - 1E-6
                                 
-#                                 JuMP.@constraint(qcm.model, [d=1:depth], sum(gates_real[(2*I-1),(2*J-1), n_g] * z_onoff_var[n_g,d] for n_g = 1:num_gates) <= v1[1])
-#                             else
-#                                 Memento.warn(_LOGGER, "Indeterminate direction for the convex-hull cut")
-#                             end
-#                             num_facets += 1
+                                JuMP.@constraint(qcm.model, [d=1:depth], sum(gates_real[(2*I-1),(2*J-1), n_g] * z_onoff_var[n_g,d] for n_g = 1:num_gates) <= v1[1])
+                            else
+                                Memento.warn(_LOGGER, "Indeterminate direction for the convex-hull cut")
+                            end
+                            num_facets += 1
                                 
-#                         end
-#                     end
-#                 end
-#             end
-#         end
+                        end
+                    end
+                end
+            end
+        end
         
-#         if num_facets > 0
-#             Memento.info(_LOGGER, "Applied $num_facets planar-hull cuts per depth of the decomposition")
-#         end
+        if num_facets > 0
+            Memento.info(_LOGGER, "Applied $num_facets planar-hull cuts per depth of the decomposition")
+        end
         
-#     end
+    end
     
-#     return
+    return
 end